Add file-backed name resolve and discovery service
